+++ lib/gem_ext/thrift/struct_union/timestamp.rb
@@ -1,5 +1,4 @@
 require 'active_support/core_ext/array/extract_options'
-require 'active_support/core_ext/object/blank'
 require 'active_support/core_ext/hash/indifferent_access'
 require 'active_support/core_ext/time/zones'
 require 'active_support/core_ext/module/delegation'