+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # yeti_thrift Changelog
 
-## v3.0.0
+## v3.0.1, October 6, 2016
+- Explicitly require `Object#try` from `ActiveSupport`.
+- New shared examples for testing required fields and unions.
+
+## v3.0.0, October 8, 2015
 - First public release.